When a zip file is being uploaded, sometimes thumbnail URLs get too long
Before this commit, MySQL would truncate the filename of thumbnail string, resulting in an inability
later to find the file (by filename).

After this commit, we check that the filename is of a length we can store in the database.
If it is not, then we use a sha1 hash of the image content as part of the image name, and
we make sure that the new auto-generated filename is short enough to fit.